automatically backed up by learn

+++ index.js
@@ -37,5 +37,5 @@ function preserveDecimal(n){
   return 2.222
 }
 function preserveDecimal(n){
-  return parseInt(n,10)
+  return parseFloat(n,10)
 }
